requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
passlib==1.7.4
pydantic-settings
sniffio==1.3.1
SQLAlchemy
typing_extensions==4.12.2
uvicorn==0.32.0
email-validator>=2.0
python-dotenv
asyncpg
PyJWT
